backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
python-slugify>=8.0.0
python-multipart>=0.0.9
PyPDF2>=3.0.1
groq>=0.5.0
